+++ bundle/internal/tf/schema/resource_pipeline.go
@@ -59,6 +59,7 @@ type ResourcePipelineClusterClusterLogConf struct {
 
 type ResourcePipelineClusterGcpAttributes struct {
 	Availability         string `json:"availability,omitempty"`
+	FirstOnDemand        int    `json:"first_on_demand,omitempty"`
 	GoogleServiceAccount string `json:"google_service_account,omitempty"`
 	LocalSsdCount        int    `json:"local_ssd_count,omitempty"`
 	ZoneId               string `json:"zone_id,omitempty"`